Enable PSDeploy functionality on Linux

+++ PSDeploy/Private/PSYaml/Private/Shadow-Copy.ps1
@@ -1,15 +1,19 @@
 function Shadow-Copy($file, $shadowPath = "$($env:TEMP)\poweryaml\shadow") {
 
-    if (-not (Test-Path $shadowPath ) ) {
-        New-Item $shadowPath -ItemType directory | Out-Null
-    }
+    if(!(Test-Path -Path Variable:\IsWindows) -or $IsWindows) {
+        if (-not (Test-Path $shadowPath ) ) {
+            New-Item $shadowPath -ItemType directory | Out-Null
+        }
 
-    try {
-        Copy-Item $file $shadowPath -Force -ErrorAction SilentlyContinue
-    } catch {
-        "Attempted to write over locked file, continuing..." | Write-Debug
+        try {
+            Copy-Item $file $shadowPath -Force -ErrorAction SilentlyContinue
+        } catch {
+            "Attempted to write over locked file, continuing..." | Write-Debug
+        }
+        $fileName = (Get-Item $file).Name
+        "$shadowPath\$fileName"
     }
-
-    $fileName = (Get-Item $file).Name
-    "$shadowPath\$fileName"
+    else {
+        $file
+    }
 }
